Can run qEngine and show results.
The results graphs are populated from console output.
Added a help dialog showing the built in functions.

+++ SQMEditor/ResultsDock.h
@@ -17,7 +17,7 @@ public:
     ~ResultsDock();
     
     void clear();
-    void addData(const QVector<int> &probs, int measure);
+    void addData(const QVector<double> &probs, int measure);
     
 private:
     QBarSeries *_probsSeries;
